+++ app/models/adverse_event.rb
@@ -90,6 +90,17 @@ class AdverseEvent < ActiveRecord::Base
       .distinct
   end
 
+  def self.generate_csv(project)
+    CSV.generate do |csv|
+      csv << csv_attributes
+      project.adverse_events.find_each { |ae| csv << ae.to_csv_array }
+    end
+  end
+
+  def to_csv_array
+    AdverseEvent.csv_attributes.collect { |csv_attr| send(csv_attr) }
+  end
+
   private
 
   # Adverse Events reports are sent to unblinded project editors
@@ -106,5 +117,18 @@ class AdverseEvent < ActiveRecord::Base
     users_to_email.each do |user_to_email|
       UserMailer.adverse_event_reported(self, user_to_email).deliver_later
     end
+  end
+
+  def self.csv_attributes
+    [:name, :reported_by, :subject_code, :reported_on, :description,
+     :serious, :closed]
+  end
+
+  def reported_by
+    user.name
+  end
+
+  def reported_on
+    adverse_event_date
   end
 end